+++ scripts/wrapper.sh
@@ -61,10 +61,10 @@ case "$1" in
       *) e "Wrapper does not support 'cargo $2'" ;;
     esac
     ;;
-  defmt-print) ensure_cargo defmt-print 1.0.0-rc.1 ;;
-  mdbook) ensure_cargo mdbook 0.5.0-alpha.1 ;;
-  nrfdfu) ensure_cargo nrfdfu 0.2.0 ;;
-  probe-rs) ensure_cargo probe-rs-tools 0.29.1 ;;
+  defmt-print) ensure_cargo defmt-print 1.0.0 ;;
+  mdbook) ensure_cargo mdbook 0.5.0-beta.1 ;;
+  nrfdfu) ensure_cargo nrfdfu 0.2.1 ;;
+  probe-rs) ensure_cargo probe-rs-tools 0.30.0 ;;
   rust-objcopy|rust-size) ensure_cargo cargo-binutils 0.4.0 ;;
   taplo) ensure_cargo taplo-cli 0.10.0 ;;
   trunk) ensure_cargo trunk 0.21.14 ;;