+++ scripts/wrapper.sh
@@ -57,12 +57,12 @@ case "$1" in
   cargo)
     case "$2" in
       bloat) ensure_cargo cargo-bloat 0.12.1 ;;
-      upgrade) ensure_cargo cargo-edit 0.13.7 ;;
+      upgrade) ensure_cargo cargo-edit 0.13.8 ;;
       *) e "Wrapper does not support 'cargo $2'" ;;
     esac
     ;;
   defmt-print) ensure_cargo defmt-print 1.0.0 ;;
-  mdbook) ensure_cargo mdbook 0.5.0-beta.1 ;;
+  mdbook) ensure_cargo mdbook 0.5.1 ;;
   nrfdfu) ensure_cargo nrfdfu 0.2.1 ;;
   probe-rs) ensure_cargo probe-rs-tools 0.30.0 ;;
   rust-objcopy|rust-size) ensure_cargo cargo-binutils 0.4.0 ;;