+++ crates/cli-tools/CHANGELOG.md
@@ -103,4 +103,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 7 -->
+<!-- Increment to skip CHANGELOG.md test: 8 -->